Simplify signing in to our demo
This way people don't have to write the username and password, and
accessing the admin area will be possible even if missing our text on
top of the page where the username and password are displayed.

In order to avoid duplication, we're getting the email from the database
and are changing the text on top of the page accordingly.