Don't ping search engines after sitemap generation
The Bing sitemap submission was removed in sitemap_generator 6.3.0 [1],
and Google has also deprecated the ping and we're getting an error when
deploying:

```
Ping failed for Google: #<OpenURI::HTTPError: 404 Sitemaps ping is
deprecated.
```

Since those were the only two search engines we were pinging in the
past, we're removing the code that did so.

[1] Pull request 408 in https://github.com/kjvarga/sitemap_generator