Ignore Sass warnings until we update our dependencies
We're getting one warning when compiling the assets due to the code we
use from font-awesome-sass, and a lot of warnings due to the code we use
from foundation.

Since these warnings are very annoying, and we get them both when
deploying and every time we change an SCSS file in development, we're
silencing them.

I haven't found the way to pass the `quiet_deps` option to the Sprockets
processor, so I'm monkey-patching the Sass engine instead.