Add Style/SingleLineDoEndBlock rubocop rule
This rule was added in Rubocop version 1.57.0. Even if we were always
following it, we think it's useful.

Note this rule doesn't detect the fact that you can still write
multiline lambdas using `{}` as delimiters instead of `do...end`, which
we fixed in the previous commit. But at least it detects the opposite
case.