Allow customizing the text to display poll dates
Since we were using `I18n.t`, our monkey-patch of the `t` helper wasn't
being applied.

Note that polls with no start or end date are invalid, so I don't think
it makes sense to write a test for the `polls.no_dates` scenario. I
wonder whether we should remove the code for this condition.